Move hooks into 'modules' directory
