+++ packages/react-native-web/src/exports/Text/index.js
@@ -15,10 +15,10 @@ import { forwardRef, useContext, useRef } from 'react';
 import createElement from '../createElement';
 import css from '../StyleSheet/css';
 import pick from '../../modules/pick';
-import useElementLayout from '../../hooks/useElementLayout';
+import useElementLayout from '../../modules/useElementLayout';
 import useMergeRefs from '../../modules/useMergeRefs';
-import usePlatformMethods from '../../hooks/usePlatformMethods';
-import useResponderEvents from '../../hooks/useResponderEvents';
+import usePlatformMethods from '../../modules/usePlatformMethods';
+import useResponderEvents from '../../modules/useResponderEvents';
 import StyleSheet from '../StyleSheet';
 import TextAncestorContext from './TextAncestorContext';
