+++ packages/react-native-web/src/modules/usePlatformMethods/index.js
@@ -9,8 +9,8 @@
 
 import type { ElementRef } from 'react';
 
-import UIManager from '../exports/UIManager';
-import createDOMProps from '../modules/createDOMProps';
+import UIManager from '../../exports/UIManager';
+import createDOMProps from '../../modules/createDOMProps';
 import { useImperativeHandle, useRef } from 'react';
 
 function setNativeProps(node, nativeProps, classList, pointerEvents, style, previousStyleRef) {
