+++ packages/react-native-web/src/modules/useResponderEvents/ResponderTouchHistoryStore.js
