+++ packages/react-native-web/src/modules/useResponderEvents/createResponderEvent.js
