+++ packages/react-native-web/src/modules/useResponderEvents/utils.js
